input_stream converted to typescript
