--- src/input/input_stream.js
@@ -1,327 +1,0 @@
-import NodeInputStream from './input_stream_node';
-import ImageLoader from './image_loader';
-
-const InputStream = {
-    createVideoStream(video) {
-        const that = {};
-        let _config = null;
-        const _eventNames = ['canrecord', 'ended'];
-        const _eventHandlers = {};
-        let _calculatedWidth;
-        let _calculatedHeight;
-        const _topRight = { x: 0, y: 0 };
-        const _canvasSize = { x: 0, y: 0 };
-
-        function initSize() {
-            const width = video.videoWidth;
-            const height = video.videoHeight;
-
-            _calculatedWidth = _config.size ? width / height > 1 ? _config.size : Math.floor((width / height) * _config.size) : width;
-            _calculatedHeight = _config.size ? width / height > 1 ? Math.floor((height / width) * _config.size) : _config.size : height;
-
-            _canvasSize.x = _calculatedWidth;
-            _canvasSize.y = _calculatedHeight;
-        }
-
-        that.getRealWidth = function () {
-            return video.videoWidth;
-        };
-
-        that.getRealHeight = function () {
-            return video.videoHeight;
-        };
-
-        that.getWidth = function () {
-            return _calculatedWidth;
-        };
-
-        that.getHeight = function () {
-            return _calculatedHeight;
-        };
-
-        that.setWidth = function (width) {
-            _calculatedWidth = width;
-        };
-
-        that.setHeight = function (height) {
-            _calculatedHeight = height;
-        };
-
-        that.setInputStream = function (config) {
-            _config = config;
-            video.src = (typeof config.src !== 'undefined') ? config.src : '';
-        };
-
-        that.ended = function () {
-            return video.ended;
-        };
-
-        that.getConfig = function () {
-            return _config;
-        };
-
-        that.setAttribute = function (name, value) {
-            video.setAttribute(name, value);
-        };
-
-        that.pause = function () {
-            video.pause();
-        };
-
-        that.play = function () {
-            video.play();
-        };
-
-        that.setCurrentTime = function (time) {
-            if (_config.type !== 'LiveStream') {
-                video.currentTime = time;
-            }
-        };
-
-        that.addEventListener = function (event, f, bool) {
-            if (_eventNames.indexOf(event) !== -1) {
-                if (!_eventHandlers[event]) {
-                    _eventHandlers[event] = [];
-                }
-                _eventHandlers[event].push(f);
-            } else {
-                video.addEventListener(event, f, bool);
-            }
-        };
-
-        that.clearEventHandlers = function () {
-            _eventNames.forEach((eventName) => {
-                const handlers = _eventHandlers[eventName];
-                if (handlers && handlers.length > 0) {
-                    handlers.forEach((handler) => {
-                        video.removeEventListener(eventName, handler);
-                    });
-                }
-            });
-        };
-
-        that.trigger = function (eventName, args) {
-            let j;
-            const handlers = _eventHandlers[eventName];
-
-            if (eventName === 'canrecord') {
-                initSize();
-            }
-            if (handlers && handlers.length > 0) {
-                for (j = 0; j < handlers.length; j++) {
-                    handlers[j].apply(that, args);
-                }
-            }
-        };
-
-        that.setTopRight = function (topRight) {
-            _topRight.x = topRight.x;
-            _topRight.y = topRight.y;
-        };
-
-        that.getTopRight = function () {
-            return _topRight;
-        };
-
-        that.setCanvasSize = function (size) {
-            _canvasSize.x = size.x;
-            _canvasSize.y = size.y;
-        };
-
-        that.getCanvasSize = function () {
-            return _canvasSize;
-        };
-
-        that.getFrame = function () {
-            return video;
-        };
-
-        return that;
-    },
-    createLiveStream(video) {
-        video.setAttribute('autoplay', true);
-        const that = InputStream.createVideoStream(video);
-
-        that.ended = function () {
-            return false;
-        };
-
-        return that;
-    },
-    createImageStream() {
-        const that = {};
-        let _config = null;
-
-        let width = 0;
-        let height = 0;
-        let frameIdx = 0;
-        let paused = true;
-        let loaded = false;
-        let imgArray = null;
-        let size = 0;
-        const offset = 1;
-        let baseUrl = null;
-        let ended = false;
-        let calculatedWidth;
-        let calculatedHeight;
-        const _eventNames = ['canrecord', 'ended'];
-        const _eventHandlers = {};
-        const _topRight = { x: 0, y: 0 };
-        const _canvasSize = { x: 0, y: 0 };
-
-        function loadImages() {
-            loaded = false;
-            ImageLoader.load(baseUrl, (imgs) => {
-                imgArray = imgs;
-                if (imgs[0].tags && imgs[0].tags.orientation) {
-                    switch (imgs[0].tags.orientation) {
-                        case 6:
-                        case 8:
-                            width = imgs[0].img.height;
-                            height = imgs[0].img.width;
-                            break;
-                        default:
-                            width = imgs[0].img.width;
-                            height = imgs[0].img.height;
-                    }
-                } else {
-                    width = imgs[0].img.width;
-                    height = imgs[0].img.height;
-                }
-                calculatedWidth = _config.size ? width / height > 1 ? _config.size : Math.floor((width / height) * _config.size) : width;
-                calculatedHeight = _config.size ? width / height > 1 ? Math.floor((height / width) * _config.size) : _config.size : height;
-                _canvasSize.x = calculatedWidth;
-                _canvasSize.y = calculatedHeight;
-                loaded = true;
-                frameIdx = 0;
-                setTimeout(() => {
-                    publishEvent('canrecord', []);
-                }, 0);
-            }, offset, size, _config.sequence);
-        }
-
-        function publishEvent(eventName, args) {
-            let j;
-            const handlers = _eventHandlers[eventName];
-
-            if (handlers && handlers.length > 0) {
-                for (j = 0; j < handlers.length; j++) {
-                    handlers[j].apply(that, args);
-                }
-            }
-        }
-
-
-        that.trigger = publishEvent;
-
-        that.getWidth = function () {
-            return calculatedWidth;
-        };
-
-        that.getHeight = function () {
-            return calculatedHeight;
-        };
-
-        that.setWidth = function (newWidth) {
-            calculatedWidth = newWidth;
-        };
-
-        that.setHeight = function (newHeight) {
-            calculatedHeight = newHeight;
-        };
-
-        that.getRealWidth = function () {
-            return width;
-        };
-
-        that.getRealHeight = function () {
-            return height;
-        };
-
-        that.setInputStream = function (stream) {
-            _config = stream;
-            if (stream.sequence === false) {
-                baseUrl = stream.src;
-                size = 1;
-            } else {
-                baseUrl = stream.src;
-                size = stream.length;
-            }
-            loadImages();
-        };
-
-        that.ended = function () {
-            return ended;
-        };
-
-        that.setAttribute = function () {};
-
-        that.getConfig = function () {
-            return _config;
-        };
-
-        that.pause = function () {
-            paused = true;
-        };
-
-        that.play = function () {
-            paused = false;
-        };
-
-        that.setCurrentTime = function (time) {
-            frameIdx = time;
-        };
-
-        that.addEventListener = function (event, f) {
-            if (_eventNames.indexOf(event) !== -1) {
-                if (!_eventHandlers[event]) {
-                    _eventHandlers[event] = [];
-                }
-                _eventHandlers[event].push(f);
-            }
-        };
-
-        that.setTopRight = function (topRight) {
-            _topRight.x = topRight.x;
-            _topRight.y = topRight.y;
-        };
-
-        that.getTopRight = function () {
-            return _topRight;
-        };
-
-        that.setCanvasSize = function (canvasSize) {
-            _canvasSize.x = canvasSize.x;
-            _canvasSize.y = canvasSize.y;
-        };
-
-        that.getCanvasSize = function () {
-            return _canvasSize;
-        };
-
-        that.getFrame = function () {
-            let frame;
-
-            if (!loaded) {
-                return null;
-            }
-            if (!paused) {
-                frame = imgArray[frameIdx];
-                if (frameIdx < (size - 1)) {
-                    frameIdx++;
-                } else {
-                    setTimeout(() => {
-                        ended = true;
-                        publishEvent('ended', []);
-                    }, 0);
-                }
-            }
-            return frame;
-        };
-
-        return that;
-    },
-};
-
-export default InputStream;
-export { NodeInputStream };
